cleanup: rename benchmark modules to the bench_ convention
Apply the naming convention (bench_ = library/engine module) to the three
modules that still carried axis-specific prefixes, preserving history via
git mv:

- arch_bench_routes      -> bench_routes
- order_bench_orderings  -> bench_order
- wording_bench_alternates -> bench_wording_alternates

Update every import across the orchestrator, trace parser, dry-run, config
parser, run_source, and the repo-root tests. Refresh the README: rename the
module tree, drop the retired configs and the broken orchestrator-CLI
examples, and document that the run_* scripts are the only entry points.
