+++ tests/phase1_prompt_benchmarking/architecture_benchmarking_workflow/bench_orchestrator.py
@@ -42,7 +42,7 @@ from mozzarellm.utils.prompt_factory import (
     make_single_cluster_analysis_user_prompt,
 )
 
-from .arch_bench_routes import Route
+from .bench_routes import Route
 from .bench_configparse import BenchmarkConfig, TimingConfig
 from .bench_dry_run import (
     _load_bundle_genes,
@@ -51,7 +51,7 @@ from .bench_dry_run import (
 )
 from .bench_metricfns import compute_all_metrics
 from .bench_reportgen import generate_report
-from .order_bench_orderings import compose_stepwise_turns_from_route
+from .bench_order import compose_stepwise_turns_from_route
 
 # =============================================================================
 # HELPERS
